* Remove unnecessary Makefile.
* created entry points.
