+++ .travis.yml
@@ -5,6 +5,8 @@ python:
   - "pypy"
 # command to install dependencies
 install:
-  - pip install nose flake8
+  - pip install nose flake8 termcolor git-pull-request
+  - pip install xmltodict gitpython rosinstall_generator
+  - pip install rosdistro requests
 # command to run tests
 script: nosetests
