+++ ebuild_generator/generators/ebuild/__init__.py
@@ -1,0 +1,3 @@
+from ebuild_generator.generators.ebuild.run import main
+if __name__ == '__main__':
+    main()
